Fixing the tests.
