+++ src/commands/scheduling.js
@@ -8,6 +8,7 @@ const Q = require("q");
 
 const slack = require('../slack.js');
 const heltour = require('../heltour.js');
+const subscription = require('../subscription.js');
 
 var EXTREMA_DEFAULTS = {
     'isoWeekday': 2,
@@ -523,18 +524,6 @@ function ambientScheduling(bot, message) {
     });
     return deferred.promise;
 }
-
-subscription.register(chesster, 'a-game-is-scheduled', function(target, context) {
-    // TODO: put these date formats somewhere, probably config?
-    var friendlyFormat = "ddd @ HH:mm";
-    target = slack.getSlackUserFromNameOrID(target);
-    var targetDate = context.result.date.clone().utcOffset(target.tz_offset/60);
-    context['yourDate'] = targetDate.format(friendlyFormat);
-    var fullFormat = "YYYY-MM-DD @ HH:mm UTC";
-    context['realDate'] = context.result.date.format(fullFormat);
-    return "{white.name} vs {black.name} in {leagueName} has been scheduled for {realDate}, which is {yourDate} for you.".format(context);
-});
-
 
 module.exports.getRoundExtrema = getRoundExtrema;
 module.exports.parseScheduling = parseScheduling;
